requirements.txt
langchain
langchain-groq
python-dotenv
langchain-community
langchain-core
streamlit